@@ -4,7 +4,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.0",
         "ext-ctype": "*",
         "ext-exif": "*",
         "ext-iconv": "*",
@@ -15,6 +15,7 @@
         "doctrine/doctrine-bundle": "^2.2",
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
+        "knplabs/knp-paginator-bundle": "^5.4",
         "liip/imagine-bundle": "^2.5",
         "nelmio/cors-bundle": "^2.1",
         "phpdocumentor/reflection-docblock": "^5.2",
@@ -27,6 +28,7 @@
         "symfony/flex": "^1.3.1",
         "symfony/form": "5.2.*",
         "symfony/framework-bundle": "5.2.*",
+        "symfony/google-mailer": "5.2.*",
         "symfony/http-client": "5.2.*",
         "symfony/intl": "5.2.*",
         "symfony/mailer": "5.2.*",
