@@ -14,9 +14,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-/**
- * @Route("/admin/user")
- */
+#[Route('/admin/user')]
 class UserController extends AbstractController
 {
     public function __construct(DatatableFactory $factory, DatatableResponse $response)
@@ -26,12 +24,12 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/", name="admin_user_index", methods={"GET"})
      * @param UserRepository $userRepository
      * @param Request $request
      * @return Response
      * @throws \Exception
      */
+    #[Route('/',name: 'admin_user_index', methods: ["GET"])]
     public function index(UserRepository $userRepository, Request $request): Response
     {
         $isAjax = $request->isXmlHttpRequest();
@@ -52,10 +50,10 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="admin_user_new", methods={"GET","POST"})
      * @param Request $request
      * @return Response
      */
+    #[Route('/new',name: 'admin_user_new', methods: ["GET","POST"])]
     public function new(Request $request): Response
     {
         $user = new User();
@@ -77,10 +75,10 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="admin_user_show", methods={"GET"})
      * @param User $user
      * @return Response
      */
+    #[Route('/{id}',name: 'admin_user_show', methods: ["GET"])]
     public function show(User $user): Response
     {
         return $this->render('Admin/user/show.html.twig', [
@@ -89,11 +87,11 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/edit", name="admin_user_edit", methods={"GET","POST"})
      * @param Request $request
      * @param User $user
      * @return Response
      */
+    #[Route('/{id}/edit',name: 'admin_user_edit', methods: ["GET","POST"])]
     public function edit(Request $request, User $user): Response
     {
         $form = $this->createForm(UserType::class, $user);
@@ -117,6 +115,7 @@ class UserController extends AbstractController
      * @param User $user
      * @return Response
      */
+    #[Route('/{id}',name: 'admin_user_delete', methods: ["DELETE"])]
     public function delete(Request $request, User $user): Response
     {
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
