@@ -4,6 +4,8 @@ namespace App\Form;
 
 use App\Entity\Comment;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -12,17 +14,20 @@ class CommentType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('content')
-            ->add('created_at')
-            ->add('enabled')
-            ->add('level')
-            ->add('picture')
+            ->add('content', TextareaType::class, [
+                "label" => " ",
+                "attr" => ["placeholder" => "form.comment.placeholder.content"]
+            ])
+            ->add("submit", SubmitType::class, [
+                "label" => "form.label.submit"
+            ])
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
+            'translation_domain' => "OurTripsTrans",
             'data_class' => Comment::class,
         ]);
     }
