@@ -1,6 +1,8 @@
 <?php
 
+
 namespace App\Form;
+
 
 use App\Entity\User;
 use Symfony\Component\Form\AbstractType;
@@ -13,7 +15,7 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class UserType extends AbstractType
+class UserEditType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
@@ -21,14 +23,6 @@ class UserType extends AbstractType
             ->add('email', EmailType::class, [
                 "label" => "form.user.label.email",
                 "attr" => ["class" => "form-case"]
-            ])
-            ->add('password', RepeatedType::class, [
-                'type' => PasswordType::class,
-                'invalid_message' => 'constraints.user.password.repeatIinvalid',
-                'options' => ['attr' => ['class' => 'password-field form-case']],
-                'required' => true,
-                'first_options'  => ['label' => 'form.user.label.password'],
-                'second_options' => ['label' => 'form.user.label.repeatPassword'],
             ])
             ->add('firstname', TextType::class, [
                 "label" => "form.user.label.firstname",
@@ -38,12 +32,12 @@ class UserType extends AbstractType
                 "label" => "form.user.label.lastname",
                 "attr" => ["class" => "form-case"]
             ])
-        ->add("submit", SubmitType::class, [
-            "label" => "form.label.submit"
-        ])
-        ->add("return", ButtonType::class, [
-            "label" => "form.user.label.returnbutton"
-        ]);
+            ->add("submit", SubmitType::class, [
+                "label" => "form.label.submit"
+            ])
+            ->add("return", ButtonType::class, [
+                "label" => "form.user.label.returnbutton"
+            ]);
     }
 
     public function configureOptions(OptionsResolver $resolver)
